#821 [REFACTOR] shared/component/input 분류 완료
