+++ src/components/Fieldset/index.js
@@ -1,7 +1,6 @@
 import CategoryButton from './CategoryButton/CategoryButton.jsx';
 import CategoryFieldset from './Fieldset/Fieldset.jsx';
 import Dropdown from './Dropdown/Dropdown.jsx';
-import Textarea from './Textarea/Textarea';
 import TextField from './TextField/TextField';
 
-export { CategoryButton, CategoryFieldset, Dropdown, Textarea, TextField };
+export { CategoryButton, CategoryFieldset, Dropdown, TextField };
