+++ src/components/Input/index.js
@@ -1,4 +1,2 @@
-export { default as Input } from './Input/Input';
-export { default as SpecialInput } from './Input/SpecialInput';
 export { default as InputItem } from './InputItem/InputItem';
 export { default as InputList } from './InputList/InputList';
