+++ src/components/index.js
@@ -14,7 +14,6 @@ export * from './Header';
 export * from './Icon';
 export * from './Input';
 export * from './InputBar';
-export * from './InputPassword';
 export * from './List';
 export * from './ListHeader';
 export * from './Loading';
