+++ src/pages/LoginPage/FindIdPage/FindIdPage.jsx
@@ -3,9 +3,8 @@ import { useNavigate } from 'react-router-dom';
 
 import { useFindId } from '@/apis';
 
+import { Input, Submit } from '@/shared/component';
 import { Icon } from '@/components/Icon';
-import { Input } from '@/components/Input';
-import { Submit } from '@/shared/component';
 
 import { FetchLoadingOverlay } from '@/components/Loading';
 import { LOADING_MESSAGE } from '@/constants';
