+++ src/pages/LoginPage/SignUpPage/SignUpPageStages/AuthorizationPage.jsx
@@ -2,8 +2,7 @@ import { useState, useEffect } from 'react';
 
 import { useSendUser, useCertifyUser } from '@/apis';
 
-import { Input } from '@/components/Input';
-import { Submit } from '@/shared/component';
+import { Input, Submit } from '@/shared/component';
 
 import styles from './AuthorizationPage.module.css';
