+++ src/shared/component/index.js
@@ -1,4 +1,11 @@
+// button
 export { default as ActionButton } from './button/ActionButton/ActionButton';
 export { default as ResetButton } from './button/ResetButton/ResetButton';
 export { default as Submit } from './button/Submit/Submit';
 export { default as WriteButton } from './button/WriteButton/WriteButton';
+
+// input
+export { default as Input } from './input/Input/Input';
+export { default as InputPassword } from './input/InputPassword/InputPassword';
+export { default as Textarea } from './input/Textarea/Textarea';
+export { default as SpecialInput } from './input/Input/SpecialInput';
