+++ src/shared/component/input/InputPassword/index.js
